% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourBlindPalette.R
\name{cbPalette}
\alias{cbPalette}
\title{Generate a colour-blind friendly palette}
\usage{
cbPalette()
}
\value{

}
\description{
Generate a colour-blind friendly palette
}
\examples{
\dontrun{
## Call BRECIplot replacing default colours with colour-blind friendly palette

BRECIplot(ras1, ras2, plotTitle = "Acacia linifolia", outFilename = "~/testPlot.png", binCols = cbPalette())
}
}
