% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gain_loss_maps.R
\name{gainLossMaps}
\alias{gainLossMaps}
\title{Plot gain and loss maps}
\usage{
gainLossMaps(
  ras1 = NULL,
  ras2 = NULL,
  brks = c(0, 0.2, 0.4, 0.6, 0.8, 1),
  binNames = c("V. low", "Low", "Med.", "High", "V. high"),
  binCols = c("dodgerblue", "darkolivegreen2", "gold1", "orange", "red"),
  mapsToMake = "both",
  outFolder = NULL,
  saveToRaster = FALSE,
  saveImages = FALSE
)
}
\arguments{
\item{ras1}{rasterLayer. Full path to the baseline raster (earliest in time).}

\item{ras2}{rasterLayer. Full path to the future raster.}

\item{brks}{Numeric. An array of bin break points. Default gives 5 equally spaced bins between 0 and 1.}

\item{binNames}{Character. An array of names assigned to the bins.}

\item{binCols}{Character. An array of standard R colour names or hexadecimal colour values.}

\item{mapsToMake}{Character. Make 'both' gain and loss maps, just 'gain' maps, or just 'loss' maps.}

\item{outFolder}{Character. Path to an output folder into which rasters will be written.}

\item{saveToRaster}{Logical. Save rasters to files in the folder specified in \emph{outFolder}? Default is FALSE.}

\item{saveImages}{Logical. Save images of rasters in png format in the folder specified in \emph{outFolder}? Default is FALSE.}
}
\description{
Generate maps for overall gain or loss and maps showing each class transition
}
\details{
{

Overall gain plot shows areas which transitioned from a lesser class to the indicated final class.

Similarly, the overall loss map shows areas which transitions from a higher class to the indicated final class.

Individual maps are produced showing each class transition which had at least one raster cell make the indicated transition.

Note that the maximum number of transition plots grows quickly with the number of suitability bins or classes. The number of possible transitions is:

\loadmathjax
\mjdeqn{ k = \sum_{i = 1}^{N - 1} N - i}{ASCII representation}

For the default number of bins, \emph{N} = 5 and so \emph{k} = 15. Naturally, if the function is called with \emph{mapsToMake} = 'both', the maximum number of transition maps could be 2 * \emph{k}.

Default values for \emph{brks}, and a matching set of bin names (\emph{binNames}) and colours (\emph{binCols}) match those for \emph{BRECIplot}, but can be changed to as desired.

If \emph{saveToFile} = FALSE (the default) then plots are made on the standard graphics device. This allows for users to review/manually save plots as images. However, setting \emph{saveToFile} suppresses plotting to the standard graphics device. File names are generated by prefixing the name with 'Gain_map_" or 'Loss_map' followed by a concatenation of the \emph{from} bin name followed by the \emph{to} bin name.

Finally, users should be aware that processing large raster files will have large impacts on processing time, and memory and disk storage requirements.
}
}
\examples{
\dontrun{ }

}
