% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRECI.R
\name{BRECIplot}
\alias{BRECIplot}
\title{Make a BRECI plot}
\usage{
BRECIplot(
  ras1 = NULL,
  ras2 = NULL,
  outFilename = NULL,
  saveToFile = TRUE,
  plotTitle = "",
  plotWidth = 275,
  plotHeight = 200,
  brks = c(0, 0.2, 0.4, 0.6, 0.8, 1),
  binNames = c("V. low", "Low", "Med.", "High", "V. high"),
  binCols = c("dodgerblue", "darkolivegreen2", "gold1", "orange", "red"),
  scaleFactor = "numCells"
)
}
\arguments{
\item{ras1}{Character. Full path to the baseline raster (earliest in time).}

\item{ras2}{Character. Full path to the future raster.}

\item{outFilename}{Character. Full path to the output file.}

\item{saveToFile}{Logical. Should the output plot be saved to file. DEFAULT = TRUE.}

\item{plotTitle}{Character. Title to be give to the plot.}

\item{plotWidth}{Integer. Width of the output plot in pixels. Default is 250 pixels.}

\item{plotHeight}{Integer. height of the output plot in pixels. Default is 188 pixels.}

\item{brks}{Numeric. An array of bin break points. Default gives 5 equally spaced bins between 0 and 1.}

\item{binNames}{Character. An array of names assigned to the bins.}

\item{binCols}{Character. An array of standard R colour names or hexadecimal colour values.}

\item{scaleFactor}{Character. Selects scaling method for bin values. One of 'ras1Bins' or 'numCells' (default). See Details for an explanation.}
}
\value{
TRUE
}
\description{
Make a BRECI plot from two rasters output by an ENM fitting process. Typically, this is MaxEnt but any raster output from any ENM fitting method is acceptable provided that the rasters are on identical geometries and have values in gridcells scaled in the same way.
}
\details{
{
The two rasters to be compared using this function MUST:
\enumerate{
 \item Have the same extent and resolution (grid cell size); and,

 \item Have values in gridcells scaled in the same way. The rasters may be
 normalised values such that the sum of grid cell values across the extent = 1
 (as in MaxEnt raw output) or adjusted so that values in a grid cell may range
 from 0 to 1 (as in MaxEnt logistic output).
}

\emph{scaleFactor} determines the way in which values in bins are computed. Selecting 'ras1Bins' computes the value for a bin as a fraction of the number of pixels in the first raster. Values in a bin can therfore be greater than +/- 1, and are \emph{greatly} exaggerated in size.

For the ith bin:

 \verb{    Bin_size(i) = [ras2_count(i) - ras1_count(i)]/ras1_count(i)}

Selecting 'numCells' computes the difference in the number of cells in that bin as a proportion of the total number of cells in the raster extent.

For the ith bin:

  \verb{    Bin_size(i) = [ras2_count(i) - ras1_count(i)]/Total_number_cells}

This function is based on code developed in April 2017 but the first approach to this kind of index was developed in 2009-2010. Code for that earlier method
is no longer available.

}
}
\examples{
data(breciRasters)

# Default plot
BRECIplot(ras1, ras2, plotTitle = "Acacia linifolia", outFilename = "~/testPlot.png")

# A colour-blind freindly palette derived from info found here:
# http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/
# last accessed 2020-03-14
colTable <- matrix(c(213,94,0,204,121,167,240,228,66,0,158,115,0,114,178), 5, 3, byrow = TRUE)/256
niceColours <- rev(rgb(colTable))
BRECIplot(ras1, ras2, plotTitle = "Acacia linifolia", outFilename = "~/testPlot.png", binCols = niceColours)

}
